Create a matrix of predictions for each of the models in a caretList

Make a matrix of predictions from a list of caret models

Usage

# S3 method for class 'caretList'
predict(object, newdata = NULL, verbose = FALSE, excluded_class_id = 1L, ...)

Arguments

object

an object of class caretList

newdata

New data for predictions. It can be NULL, but this is ill-advised.

verbose

Logical. If FALSE no progress bar is printed if TRUE a progress bar is shown. Default FALSE.

excluded_class_id

Integer. The class id to drop when predicting for multiclass

...

Other arguments to pass to predict.train